+++ heeate/pages/index.tsx
@@ -1,0 +1,11 @@
+import type { NextPage } from "next";
+import Head from "next/head";
+import Image from "next/image";
+import styles from "../styles/Home.module.css";
+import App from "./todo";
+
+const Home: NextPage = () => {
+  return <App />;
+};
+
+export default Home;
